fix: Include n_cos & n_rows info in SparseMatrix
This prevents having issues as shown in the tests now failing for MLE
conversions. The issue is that empty rows/columns at the end of the
matrixes cause the algorithms to actually fail. As we have a sense of
the matrix dimensions only when we are in the context of CCS. But we
don't when we're outside of it.